+++ src/Form/formControlValue.js
@@ -7,6 +7,8 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+import update from 'react-update'
+
 export default (form, formItem) => {
   const name = formItem.props.name
   return {
@@ -18,15 +20,21 @@ export default (form, formItem) => {
         if (data instanceof Array) {
           return data[form.multipleMap[formItem.uuid]]
         }
+        return null
       }
     },
     set(value) {
+      const formData = form.state.data
       let nextData
       if (!formItem.props.multiple) {
-        nextData = form.update('set', ['data', name], value)
+        nextData = update(formData, 'set', name, value)
       } else {
-        nextData = form.update('set', ['data', name, form.multipleMap[formItem.uuid]], value)
+        if (!formData[name]) {
+          formData[name] = []
+        }
+        nextData = update(formData, 'set', [name, form.multipleMap[formItem.uuid]], value)
       }
+      form.setState({data: nextData})
       form.props.onChange && form.props.onChange(nextData)
     }
   }